apps/4/requirement.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
jiter==0.5.0
numpy==1.24.4
openai==1.43.0
pandas==2.0.3
pydantic==2.8.2
pydantic-core==2.20.1
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.7.24
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tiktoken==0.7.0
tqdm==4.66.5
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
